[MODIFY] app파일, server파일 분리
