+++ .env.sample
@@ -1,15 +1,15 @@
-DATABASE_URL = "rdbms://db_username:password@db_host:db_port/database_name"
+DATABASE_URL="rdbms://db_username:password@db_host:db_port/database_name"
 
-PORT = YOUT PORT NUMBER SHOULD BE HERE!
+PORT=YOUT PORT NUMBER SHOULD BE HERE!
 
-DB_CONNECTION = rdbms
-DB_HOST = db_host
-DB_USERNAME = db_username
-DB_PASSWORD = myPassword
-DB_DATABASE = myDB
-DB_PORT = db_port
-DB_LOGGING = TRUE
+DB_CONNECTION=rdbms
+DB_HOST=db_host
+DB_USERNAME=db_username
+DB_PASSWORD=myPassword
+DB_DATABASE=myDB
+DB_PORT=db_port
+DB_LOGGING=TRUE
 
-ALGORITHM = algorithm
-JWT_SECRET= jwt-secret
-JWT_EXPIRES_IN = date
+ALGORITHM=algorithm
+JWT_SECRET=jwt-secret
+JWT_EXPIRES_IN=date
